pyproject.toml
[tool.poetry]
name = "agroatom"
version = "0.1.0"
description = ""
authors = [

]
maintainers = [

]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
pydantic = {version = "^1.8.2", extras = ["dotenv"]}
yarl = "^1.6.3"
ujson = "^4.2.0"
tortoise-orm = "^0.17.7"
aerich = "^0.5.8"
asyncpg = "^0.24.0"
aioredis = {version = "^2.0.0", extras = ["hiredis"]}
httptools = "^0.3.0"
async-exit-stack = "^1.0.1"
async-generator = "^1.10"
python-multipart = "^0.0.5"
aioboto3 = "^8.2.0"
wsproto = "^1.0.0"
orjson = "^3.6.4"
APScheduler = "^3.8.1"
async-to-sync = "^0.2.2"
dramatiq = {version = "^1.12.0", extras = ["rabbitmq", "watch]"]}
typer = "^0.4.0"
rq = "^1.10.0"
pykml = "^0.2.0"
pandas = "^1.3.4"
opencv-python-headless = "^4.5.4"
torch = "^1.10.0"
Pillow = "^8.4.0"
torchvision = "^0.11.1"
matplotlib = "^3.4.3"
tqdm = "^4.62.3"
seaborn = "^0.11.2"

[tool.poetry.dev-dependencies]
pytest = "^6.0"
flake8 = "^3.9.2"
mypy = "^0.910"
isort = "^5.9.3"
yesqa = "^1.2.3"
pre-commit = "^2.11.0"
wemake-python-styleguide = "^0.15.3"
black = "==21.7b0"
autoflake = "^1.4"
pytest-cov = "^2.12.1"
pytest-asyncio = "^0.15.1"
nest-asyncio = "^1.5.1"
pytest-env = "^0.6.2"
fakeredis = "^1.6.1"
requests = "^2.26.0"
asynctest = "^0.13.0"


[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["agroatom",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_return_any = false

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
]
env = [
    "AGROATOM_DB_BASE=agroatom_test",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
